requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Cache
flask-assets
flask-redis
Flask-MustacheJS
celery
ensure
pyscss
pytest-flask
Jinja2
MarkupSafe
SQLAlchemy
Werkzeug
aniso8601
argparse
itsdangerous
pbkdf2
python-dateutil
pytz
pystache
six
wsgiref
jsmin
boto3
tzlocal
docker-py
git+ssh://git@github.com/ThreatResponse/margaritashotgun.git@master
git+ssh://git@github.com/ThreatResponse/aws_ir.git@master
git+ssh://git@github.com/ThreatResponse/ThreatPrep.git@master
